Payment_Service/payment_service/pyproject.toml
[tool.poetry]
name = "payment-service"
version = "0.1.0"
description = ""
authors = ["Hamadalishah <hammadalis459@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.12"
sqlmodel = "^0.0.22"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
python-jose = "^3.3.0"
python-dotenv = "^1.0.1"
aiokafka = "^0.11.0"
protobuf = "^5.28.2"
types-protobuf = "^5.28.0.20240924"
psycopg2-binary = "^2.9.10"
dapr = "^1.14.0"
requests = "^2.32.3"
httpx = "^0.27.2"
dependencies = "^7.7.1"
multipart = "^1.2.1"
python-multipart = "^0.0.18"
strip = "^0.0.1.dev14"
stripe = "^11.3.0"
types-stripe = "^3.5.2.20240106"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"





[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
plugins = ["mypy_stripe.plugins.StripePlugin"]
